Allow override of bullet face
Sometimes bullets don't look good in the 'org-level-x' face and so this
change allows an override face to be specified with
'org-bullets-face-name'. Set this to the face name as a string to force
a particular face to be used for bullets, or set it to nil to use the
normal Org-mode heading face.