Add option to make the invisible leading stars.
Add option to make them invisible, not hidden by means of it's colour.
Thus, the end result looks like

x Heading level 1
o Heading level 2
> Heading level 3
o Heading level 2

instead of

x Heading level 1
 o Heading level 2
  > Heading level 3
 o Heading level 2

if `org-bullets-invisible-leading-stars' is `t'. Useful if the the
contents of `org-bullets-bullet-list' makes obvious the level of the
heading, and indentation is superfluous.